+++ plugin-build/src/test/kotlin/io/sentry/android/gradle/integration/SentryPluginConfigurationCacheTest.kt
@@ -150,7 +150,7 @@ class SentryPluginConfigurationCacheTest :
               autoUploadProguardMapping = false
               autoInstallation.enabled = false
               includeDependenciesReport = false
-              telemetry = false
+              telemetry = true
             }
             """.trimIndent()
         )